// src/app/shared/models/hbc-customer/hbc-customer.model.ts
export class CustomerModel {
    name: string;
    website: string;
    email: string;
    phone: string;
    taxNumber: string;
    stockCode: string;
    startWorkingFromDate: number;
    type: string;
    address: string;
    assignedEmployeeID: number;
    customerContactId: number;
    business: string;
    customerGroup: string;
    classify: string;
    evaluate: number;
    revenueDueYear: number;
    establishmentDate: number;
    description: string;
    isPosted: true;
    role: string;
    manageDepartmentName: string;
    manageDepartmentPhone: string;
    district: string;
    city: string;
    country: string;
  }